1. Удаление FSEntryMenu 2. Унификация ListMenu 3. Добаление класса Brush и ListMenuStyle

+++ trunk/gcc/Sources/LampFM/actionlib/ActiveArea.c
@@ -195,7 +195,7 @@ int activeAreaPopItem(ActiveArea *field, ActiveAreaItem *_item)
             goto __found;
     }
 
-    printf("activeAreaPopItem(%X): Not found\n", _item);
+    printf("activeAreaPopItem(%X): Not found\n", (unsigned int)_item);
     return -1;
 
 __found:
